+++ src/main/java/com/dev/core/service/implementation/FlightServiceImpl.java
@@ -6,7 +6,10 @@ import com.dev.core.entity.FlightStatus;
 import com.dev.core.repository.FlightRepository;
 import com.dev.core.service.FlightService;
 import java.util.List;
+import java.util.Optional;
+import org.springframework.stereotype.Service;
 
+@Service
 public class FlightServiceImpl implements FlightService {
     
     private final FlightRepository flightRepository;
@@ -47,11 +50,8 @@ public class FlightServiceImpl implements FlightService {
     
     @Override
     public void remove(Flight flight) {
-        flightRepository.delete(flight);
-    }
-    
-    @Override
-    public void update(Flight flight) {
-        flightRepository.save(flight);
+        if (flightRepository.findById(flight.getId()).isPresent()) {
+            flightRepository.delete(flight);
+        }
     }
 }
